+++ lib/db.dart
@@ -33,7 +33,7 @@ const Symbol _dbKey = #gcloud.db;
 ///
 /// Accessing this getter outside of a service scope will result in an error.
 /// See the `package:gcloud/service_scope.dart` library for more information.
-DatastoreDB get dbService => ss.lookup(_dbKey);
+DatastoreDB get dbService => ss.lookup(_dbKey) as DatastoreDB;
 
 /// Registers the [DatastoreDB] object within the current service scope.
 ///